4001 - Move parseTable file
