+++ src/client/pages/Section/DataTable/Table/hooks/_parseTable.ts
